oldCode/backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Configuration and validation
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# LangChain and LLM
langchain-ollama==0.0.2
langchain-core==0.1.10
langgraph==0.0.20

# Vector store and embeddings
faiss-cpu==1.7.4
sentence-transformers==2.2.2

# PyTorch - CRITICAL: Install torch and torchvision TOGETHER from PyTorch index
# DO NOT install from PyPI - they must come from the same source
# Run this FIRST before installing other dependencies:
#   pip install torch==2.1.0 torchvision==0.16.0 --index-url https://download.pytorch.org/whl/cpu
# Or for GPU (CUDA 11.8):
#   pip install torch==2.1.0 torchvision==0.16.0 --index-url https://download.pytorch.org/whl/cu118
# These are listed here for reference only - install manually as shown above
# torch==2.1.0
# torchvision==0.16.0

numpy>=1.24.0
packaging>=21.0

pypdf==3.17.4

# Web search
tavily-python==0.3.0

# HTTP client
httpx==0.25.2

# Logging (standard library, but keeping for reference)
# python-json-logger==2.0.7  # Optional: for advanced JSON logging
PyJWT==2.8.0
